Feature/seguridad course

+++ src/main/kotlin/ar/edu/unsam/pds/services/CoursesService.kt
@@ -4,6 +4,7 @@ import ar.edu.unsam.pds.dto.request.CourseRequestDto
 import ar.edu.unsam.pds.dto.response.CourseDetailResponseDto
 import ar.edu.unsam.pds.dto.response.CourseResponseDto
 import ar.edu.unsam.pds.dto.response.CourseStatsResponseDto
+import ar.edu.unsam.pds.exceptions.InternalServerError
 import ar.edu.unsam.pds.exceptions.NotFoundException
 import ar.edu.unsam.pds.exceptions.ValidationException
 import ar.edu.unsam.pds.models.Course
@@ -15,8 +16,9 @@ import java.util.*
 
 @Service
 class CoursesService(
-    private val courseRepository: CourseRepository
-) {
+    private val courseRepository: CourseRepository,
+    private val userService: UserService
+    ) {
 
     fun getAll(query: String): List<CourseResponseDto> {
         val courses = courseRepository.getAllBy(query)
@@ -31,6 +33,7 @@ class CoursesService(
     @Transactional
     fun deleteCourse(idCourse: String) {
         val course = findCourseById(idCourse)
+        checkAdminPermissions()
 
         if (course.assignments.any { it.hasAnySubscribedUser() }) {
             throw ValidationException("No se puede eliminar un curso con usuarios inscriptos")
@@ -40,6 +43,8 @@ class CoursesService(
 
     @Transactional
     fun deleteAllById(courseIds: List<String>) {
+        checkAdminPermissions()
+
         courseIds.forEach { id ->
             val course = findCourseById(id)
             courseRepository.delete(course)
@@ -55,6 +60,8 @@ class CoursesService(
 
     @Transactional
     fun createCourse(course: CourseRequestDto): CourseResponseDto? {
+        checkAdminPermissions()
+
         val newCourse = Course(
             course.title,
             course.description,
@@ -70,4 +77,20 @@ class CoursesService(
         return Mapper.buildCourseStatsDto(course)
 
     }
+
+
+    private fun checkAdminPermissions() {
+        try {
+            val currentUser = userService.getCurrentUser()
+            if (!currentUser.isAdmin) {
+                throw NotFoundException("No tienes permiso para realizar esta acción.")
+            }
+        } catch (e: NotFoundException) {
+            throw NotFoundException("Usuario no encontrado. No tienes permiso para realizar esta acción.")
+        } catch (e: Exception) {
+            throw InternalServerError("Error al verificar los permisos. Inténtelo de nuevo más tarde.")
+        }
+    }
+
+
 }
